Create a copy for unadressable values.
The call to .Interface() requires an addressable value.  However, the
values that come out of a channel are not necessarily addressable if
they are non-pointer values.  A value is non-addressable if it doesn't
have a stable address (i.e. there is a guarantee that there will be
something meaningful at the address).  This can be interpreted as "an
addressable value is heap allocated".  So, our non-addressable value
is stack allocated.  We can turn it into an addressable value, but
copying it to the heap, which is what this code does.